+++ packages/api-reference/src/components/GettingStarted.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
+import { ScalarButton } from '@scalar/components'
 import { type ThemeId } from '@scalar/themes'
 import { ref, watch } from 'vue'
 
 import coinmarketcap from '../coinmarketcapv3.json'
 import petstore from '../petstorev3.json'
 import { type GettingStartedExamples } from '../types'
-import FlowButton from './FlowButton.vue'
 
 const props = defineProps<{
   theme: ThemeId
@@ -90,13 +90,17 @@ watch(
         Beautiful API references.
       </p>
       <div class="start-cta">
-        <FlowButton
-          label="Test Petstore"
-          @click="example = 'Petstore'" />
-        <FlowButton
-          label="Upload File"
+        <ScalarButton
+          isFullWidth
+          @click="example = 'Petstore'">
+          Test Petstore
+        </ScalarButton>
+        <ScalarButton
+          isFullWidth
           variant="outlined"
-          @click="$emit('openSwaggerEditor', 'uploadFile')" />
+          @click="$emit('openSwaggerEditor', 'uploadFile')">
+          Upload File
+        </ScalarButton>
       </div>
     </div>
     <div class="start-row">