+++ packages/api-reference/vite.standalone.config.ts
@@ -45,23 +45,24 @@ export default defineConfig({
       },
     },
   },
+  resolve: {
+    alias: [
+      // Resolve the uncompiled source code for all @scalar packages
+      // It’s working with the alias, too. It’s just required to enable HMR.
+      // It also does not match components since we want the built version
+      {
+        // Resolve the uncompiled source code for all @scalar packages
+        // @scalar/* -> packages/*/
+        // (not @scalar/*/style.css)
+        find: /^@scalar\/(?!(components\/style\.css|components\b))(.+)/,
+        replacement: path.resolve(__dirname, '../$2/src/index.ts'),
+      },
+    ],
+  },
   test: {
     coverage: {
       enabled: true,
       reporter: 'text',
     },
-  },
-  resolve: {
-    alias: [
-      // Resolve the uncompiled source code for all @scalar packages
-      // It’s working with the alias, too. It’s just required to enable HMR.
-      {
-        // Resolve the uncompiled source code for all @scalar packages
-        // @scalar/* -> packages/*/
-        // (not @scalar/*/style.css)
-        find: /^@scalar\/([^/]+)$/,
-        replacement: path.resolve(__dirname, '../$1/src/index.ts'),
-      },
-    ],
   },
 })