+++ packages/components/src/components/ScalarIconButton/ScalarIconButton.stories.ts
@@ -1,13 +1,14 @@
 import type { Meta, StoryObj } from '@storybook/vue3'
 
-import { iconNames } from '../ScalarIcon/icons/iconNames'
+import { ICONS } from '../ScalarIcon/icons/icons'
 import ScalarIconButton from './ScalarIconButton.vue'
 
 const meta = {
   component: ScalarIconButton,
   tags: ['autodocs'],
   argTypes: {
-    icon: { control: 'select', options: iconNames },
+    icon: { control: 'select', options: ICONS },
+    size: { control: 'select', options: ['xs', 'sm', 'md'] },
     label: { control: 'string' },
     variant: {
       control: 'select',
@@ -28,4 +29,8 @@ type Story = StoryObj<typeof meta>
 
 export const Base: Story = {
   args: { icon: 'Logo', label: 'Logo button' },
+}
+
+export const Disabled: Story = {
+  args: { icon: 'Logo', label: 'Logo button', disabled: true },
 }