+++ packages/components/src/components/ScalarModal/ScalarModal.vue
@@ -10,6 +10,8 @@ import { cva, cx } from '@/cva'
 import { reactive } from 'vue'
 import type { VariantProps } from 'cva'
 
+type ModalVariants = VariantProps<typeof modal>
+
 withDefaults(
   defineProps<{
     state: ReturnType<typeof useModal>
@@ -42,8 +44,6 @@ const modal = cva({
     },
   },
 })
-
-type ModalVariants = VariantProps<typeof modal>
 </script>
 
 <script lang="ts">
@@ -75,7 +75,7 @@ export const useModal = () =>
           {{ title }}
         </DialogTitle>
         <DialogDescription
-          class="scalar-modal-body relative max-h-[calc(100dvh-240px)] overflow-y-auto rounded-lg bg-back-1 py-6 pb-4 pt-3"
+          class="scalar-modal-body relative max-h-[calc(100dvh-240px)] overflow-y-auto rounded-lg bg-back-1 px-6 pb-4 pt-6"
           :class="
             cx(
               bodyClass,