+++ packages/swagger-editor/vite.config.ts
@@ -38,12 +38,13 @@ export default defineConfig({
     alias: [
       // Resolve the uncompiled source code for all @scalar packages
       // It’s working with the alias, too. It’s just required to enable HMR.
+      // It also does not match components since we want the built version
       {
         // Resolve the uncompiled source code for all @scalar packages
         // @scalar/* -> packages/*/
         // (not @scalar/*/style.css)
-        find: /^@scalar\/([^/]+)$/,
-        replacement: path.resolve(__dirname, '../$1/src/index.ts'),
+        find: /^@scalar\/(?!(components\/style\.css|components\b))(.+)/,
+        replacement: path.resolve(__dirname, '../$2/src/index.ts'),
       },
     ],
   },