+++ packages/hono-api-reference/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @scalar/hono-api-reference
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [648e5a98]
+  - @scalar/api-reference@1.9.2
+
+## 0.2.7
+
+### Patch Changes
+
+- edb02e76: fix: windows select menu not updating on prefers color scheme
+- Updated dependencies [edb02e76]
+  - @scalar/api-reference@1.9.1
+
+## 0.2.6
+
+### Patch Changes
+
+- d7da3147: fix: remove duplicate slash in example and client
+- Updated dependencies [d7da3147]
+- Updated dependencies [c45ae5e8]
+  - @scalar/api-reference@1.9.0
+
 ## 0.2.5
 
 ### Patch Changes